Reverting test dependency changes
