+++ config-builder/build.gradle
@@ -20,12 +20,3 @@ dependencies {
     compile 'org.reflections:reflections:0.9.11'
     runtime 'org.slf4j:slf4j-simple:1.7.25'
 }
-
-task testJar(type: Jar) {
-    classifier = 'test'
-    from sourceSets.test.output
-}
-
-artifacts {
-    testRuntime testJar
-}
