+++ gateway-export-plugin/build.gradle
@@ -28,7 +28,6 @@ dependencies {
     compile group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.9.6'
 
     testCompile gradleTestKit()
-    testCompile project(path: ':config-builder', configuration: 'testRuntime')
 }
 
 gradlePlugin {
