+++ gateway-export-plugin/src/test/java/com/ca/apim/gateway/cagatewayexport/tasks/explode/filter/entityfilters/ServiceFilterTest.java
@@ -9,8 +9,8 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
-import static com.ca.apim.gateway.cagatewayconfig.util.TestUtils.createFolder;
-import static com.ca.apim.gateway.cagatewayconfig.util.TestUtils.createService;
+import static com.ca.apim.gateway.cagatewayexport.util.TestUtils.createFolder;
+import static com.ca.apim.gateway.cagatewayexport.util.TestUtils.createService;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
